% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bubbleForceInfobox.R
\name{bubbleInfoboxOutput}
\alias{bubbleInfoboxOutput}
\title{Widget output function for use in Shiny}
\usage{
bubbleInfoboxOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

