% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bubbles.R
\name{bubblesOutput}
\alias{bubblesOutput}
\title{Widget output function for use in Shiny}
\usage{
bubblesOutput(outputId, width = "100\%", height = "500px")
}
\description{
Widget output function for use in Shiny
}

